Changed, merged Level, Score manager into GameManager.
