--- Assets/Prefabs/LevelManager.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 1675dddce122563478ee18d775d88ffa
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
