--- Assets/Scripts/LevelManager.cs
@@ -1,28 +1,0 @@
-using System.Threading.Tasks;
-using TMPro;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-using UnityEngine.UI;
-
-public class LevelManager : MonoBehaviour
-{
-    public static LevelManager Instance;
-
-    public void Awake()
-    {
-        if (Instance == null)
-        {
-            Instance = this;
-            DontDestroyOnLoad(gameObject);
-        }
-        else
-        {
-            Destroy(gameObject);
-        }
-    }
-    public void LoadScene(string sceneName)
-    {
-        var scene = SceneManager.LoadSceneAsync(sceneName);
-        scene.allowSceneActivation = true;
-    }
-}
