--- Assets/Scripts/LevelManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 02a7612b84675164bb66ce1eef52dc74
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
