--- Assets/Scripts/ScoreManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 971f7b8ee9af56343847fa071844c10e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
